requirements.txt
numpy==1.19.5
Flask
flask_cors
pytube
rq
redis
pydub
torch
openai
openai-whisper
python-dotenv
vaderSentiment
pytest